pyproject.toml
[tool.poetry]
name = "Messaging API"
version = "0.1.0"
description = ""
authors = ["Max Shapira <max@winoutt.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
psycopg = { extras = ["binary"], version = "3.1.9" }
pydantic-settings = { extras = ["dotenv"], version = "^2.0.2" }
djangorestframework = "^3.14.0"
whitenoise = { extras = ["brotli"], version = "^6.5.0" }

[tool.poetry.group.dev.dependencies]
black = { version = "^23.7.0", allow-prereleases = true }
ruff = "^0.3.4"
isort = "^5.12.0"
mypy = "^1.3.0"
django-stubs = { extras = ["compatible-mypy"], version = "^4.2.3" }
pytest = "7.4.0"
pytest-sugar = "^0.9.6"
djangorestframework-stubs = { extras = [
  "compatible-mypy",
], version = "^3.14.2" }

[tool.ruff]
line-length = 88
exclude = ['.git', '__pycache__', '__init__.py', '.mypy_cache', '.pytest_cache']
select = [
  "E", # pycodestyle errors
  "W", # pycodestyle warnings
  "F", # pyflakes
  "C", # flake8-comprehensions
  "B", # flake8-bugbear
]
ignore = [
  "E501", # line too long, handled by black
  "C901", # too complex
]

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88
profile = "black"

[tool.mypy]
follow_imports = "silent"
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
no_implicit_reexport = true
check_untyped_defs = true
strict_equality = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
show_error_codes = true
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[tool.django-stubs]
django_settings_module = "config.settings"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
